improve api for io::stream and io::system

+++ src/clover_lib/clover/io/stream.hpp
@@ -25,7 +25,7 @@ struct stream {
         int chan_count_out;
 
         int sample_rate;
-        float latency_ms;
+        double latency_ms;
     };
 
     stream();
